requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
html2text==2020.1.16
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
